+++ database/src/controllers/index.ts
@@ -1,0 +1,9 @@
+import createModel from "./createModel";
+import getModelById from "./getModelById";
+import getModels from "./getModels";
+
+export default {
+  getModels: getModels,
+  getModelById: getModelById,
+  createModel: createModel,
+};
